+++ lib/styles/borders.dart
@@ -2,5 +2,6 @@ import 'package:flutter/material.dart';
 
 // Widgets
 const BorderRadius borderRadiusDebugButton = BorderRadius.all(Radius.circular(6));
+const BorderRadius borderRadiusSoundModePill = BorderRadius.all(Radius.circular(6));
 const BorderRadius borderRadiusUpdateBadge = BorderRadius.all(Radius.circular(6));
 const BorderRadius borderRadiusXpBar = BorderRadius.all(Radius.circular(4));